[04장] 머스테치로 화면 구성하기 - 전체 조회 화면 만들기
