--- 김가경/freelec-springboot2-webservice/src/main/java/com/gagyeong/book/springboot/domain/posts/PostRepository.java
@@ -1,6 +1,0 @@
-package com.gagyeong.book.springboot.domain.posts;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostRepository extends JpaRepository<Posts, Long> {
-}
