+++ 김가경/freelec-springboot2-webservice/src/main/java/com/gagyeong/book/springboot/domain/posts/PostsRepository.java
@@ -1,6 +1,12 @@
 package com.gagyeong.book.springboot.domain.posts;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+
+import java.util.List;
 
 public interface PostsRepository extends JpaRepository<Posts, Long> {
+
+    @Query("SELECT p FROM Posts p ORDER BY p.id DESC")
+    List<Posts> findAllDesc();
 }
