+++ 김가경/freelec-springboot2-webservice/src/main/java/com/gagyeong/book/springboot/service/posts/PostsService.java
@@ -2,12 +2,16 @@ package com.gagyeong.book.springboot.service.posts;
 
 import com.gagyeong.book.springboot.domain.posts.Posts;
 import com.gagyeong.book.springboot.domain.posts.PostsRepository;
+import com.gagyeong.book.springboot.web.dto.PostsListResponseDto;
 import com.gagyeong.book.springboot.web.dto.PostsResponseDto;
 import com.gagyeong.book.springboot.web.dto.PostsSaveRequestDto;
 import com.gagyeong.book.springboot.web.dto.PostsUpdateRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 @Service
@@ -31,5 +35,10 @@ public class PostsService {
         Posts entity = postsRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다. id="+ id));
 
         return new PostsResponseDto(entity);
+    }
+
+    @Transactional(readOnly = true)
+    public List<PostsListResponseDto> findAllDesc() {
+        return  postsRepository.findAllDesc().stream().map(PostsListResponseDto::new).collect(Collectors.toList());
     }
 }
