+++ 김가경/freelec-springboot2-webservice/src/main/java/com/gagyeong/book/springboot/web/IndexController.java
@@ -1,0 +1,24 @@
+package com.gagyeong.book.springboot.web;
+
+import com.gagyeong.book.springboot.service.posts.PostsService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@RequiredArgsConstructor
+@Controller
+public class IndexController {
+    private final PostsService postsService;
+
+    @GetMapping("/")
+    public String index(Model model) {
+        model.addAttribute("posts", postsService.findAllDesc());
+        return "index";
+    }
+
+    @GetMapping("/posts/save")
+    public String postsSave() {
+        return "posts-save";
+    }
+}
