--- 김가경/freelec-springboot2-webservice/src/main/java/com/gagyeong/book/springboot/web/dto/IndexController.java
@@ -1,19 +1,0 @@
-package com.gagyeong.book.springboot.web.dto;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@RequiredArgsConstructor
-@Controller
-public class IndexController {
-    @GetMapping("/")
-    public String index() {
-        return "index";
-    }
-
-    @GetMapping("/posts/save")
-    public String postsSave() {
-        return "posts-save";
-    }
-}
