+++ 김가경/freelec-springboot2-webservice/src/main/java/com/gagyeong/book/springboot/web/dto/PostsListResponseDto.java
@@ -1,0 +1,21 @@
+package com.gagyeong.book.springboot.web.dto;
+
+import com.gagyeong.book.springboot.domain.posts.Posts;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+public class PostsListResponseDto {
+    private Long id;
+    private String title;
+    private String author;
+    private LocalDateTime modifiedDate;
+
+    public PostsListResponseDto(Posts entity) {
+        this.id = entity.getId();
+        this.title = entity.getTitle();
+        this.author = entity.getAuthor();
+        this.modifiedDate = entity.getModifiedDate();
+    }
+}
