--- 김가경/freelec-springboot2-webservice/src/test/java/com/gagyeong/book/springboot/Application.java
@@ -1,10 +1,0 @@
-package com.gagyeong.book.springboot;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-}
